+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { GlobalStyle } from "./styles/GlobalStyle";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Layout from "./components/Layout";
+import Layout from "./components/Layout/Layout";
 import ChatRoom from "components/ChatRoom/ChatRoom";
 
 function App() {
