+++ src/components/Layout/Layout.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { Outlet } from "react-router-dom";
 import Header from "./Header";
-import Footer from "components/Footer";
+import Footer from "components/Layout/Footer";
 
 function Layout() {
   return (
